requirements.txt
Flask==3.1.1
Flask_Cors==6.0.1
Flask[async]==3.1.1
numpy==1.26.4
langchain_core==0.3.81
langchain_community==0.3.31
langchain==0.3.27
langchain_openai==0.3.35
python-dotenv==1.0.1
pyyaml==6.0.2
waitress==3.0.2
openai==2.6.1
setuptools==80.9.0
httpx==0.27.2
pytest==8.3.5
pytest-cov==6.0.0
pytest-asyncio==0.25.3
pydantic==2.9.2
prisma==0.15.0
pyjwt==2.10.1
cryptography==44.0.3
spacy==3.8.2
presidio_analyzer==2.2.360
presidio_anonymizer==2.2.360
docxtpl==0.19.1
mammoth==1.11.0
pymupdf==1.25.5
pymupdf4llm==0.0.19
azure-keyvault-keys==4.10.0
azure-ai-ml==1.29.0
azure-appconfiguration==1.7.1
azure-core==1.38.0
azure-identity==1.23.1
featuremanagement==2.1.0
azure-appconfiguration-provider==2.2.0
